streaming/prediction/requirements.txt
pandas==1.5.3
scikit-learn==1.2.2
mlflow==2.3.2
boto3
dagshub
python-dotenv
numpy